Simplify Eleventy config file
